добавлены/замаплены/тесты News, NewsComment

+++ dao/src/main/java/by/itacademy/entity/HeroTopicMessage.java
@@ -17,7 +17,7 @@ public class HeroTopicMessage extends Message{
     @Getter
     @Setter
     @Column(name = "likes")
-    private int likeFromOthersUsers;
+    private long likeFromOthersUsers;
 
     @Getter
     @Setter
